requirements.txt
# --- AI & GRAPH ---
langchain>=1.1.3
langgraph>=1.0.5
langgraph-checkpoint-postgres>=3.0.2
langchain-core>=1.2.0
langchain-google-genai>=4.0.0
langchain-openai>=1.1.3
langchain-chroma>=1.1.0
langchain-text-splitters>=1.1.0
langfuse>=3.10.7
tavily>=1.1.0
langchain_deepseek

# --- WEB & API ---
fastapi>=0.127.0
uvicorn[standard]>=0.38.0
streamlit>=1.52.1
python-dotenv>=1.0.1
dotenv>=0.9.9

# --- DATABASE & AUTH ---
sqlalchemy>=2.0.45
psycopg[binary,pool]>=3.3.2
aiosqlite>=0.22.0
fastapi-users[sqlalchemy]>=15.0.3
asyncpg>=0.31.0

# --- UTILS ---
pydantic>=2.9.0
PyYAML>=6.0.2
typing-extensions>=4.12.0
requests~=2.32.5